# src/WC.Service.EmailDomains.API/WC.Service.EmailDomains.API.swagger.json
{
  "x-generator": "NSwag v14.2.0.0 (NJsonSchema v11.1.0.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "NSwag.AspNetCore.Launcher",
    "version": "v1"
  },
  "paths": {
    "/api/v1/email-domains": {
      "get": {
        "tags": [
          "EmailDomain"
        ],
        "summary": "Retrieves a list of email domains.",
        "operationId": "EmailDomainGet",
        "parameters": [
          {
            "name": "Filters",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "Sorts",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 2147483647.0,
              "minimum": 1.0,
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 2147483647.0,
              "minimum": 1.0,
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "withIncludes",
            "in": "query",
            "description": "Specifies whether related entities should be included in the query.",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "x-position": 5
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmailDomainDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "EmailDomain"
        ],
        "summary": "Creates new email domain.",
        "operationId": "EmailDomainCreate",
        "requestBody": {
          "x-name": "payload",
          "description": "The email domain content.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailDomainCreateDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "The result of creation. CreateActionResultDto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateActionResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "Admin"
            ]
          }
        ]
      }
    },
    "/api/v1/email-domains/{id}": {
      "get": {
        "tags": [
          "EmailDomain"
        ],
        "summary": "Retrieves a email domain by its ID.",
        "operationId": "EmailDomainGetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the email domain to retrieve.",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailDomainDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "Admin"
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "EmailDomain"
        ],
        "summary": "Updates a email domain by ID.",
        "operationId": "EmailDomainUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the email domain to update.",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "patchDocument",
          "description": "The JSON patch document containing updates.",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "Admin"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "EmailDomain"
        ],
        "summary": "Deletes a email domain by ID.",
        "operationId": "EmailDomainDelete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the email domain to delete.",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "Admin"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "EmailDomainDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DtoBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "domainName"
            ],
            "properties": {
              "domainName": {
                "type": "string",
                "minLength": 1
              }
            }
          }
        ]
      },
      "DtoBase": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          }
        }
      },
      "ErrorDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "stackTrace": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "extensions": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
      "CreateActionResultDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          }
        }
      },
      "EmailDomainCreateDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "domainName"
        ],
        "properties": {
          "domainName": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "Operation": {
        "allOf": [
          {
            "$ref": "#/components/schemas/OperationBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "value": {
                "nullable": true
              }
            }
          }
        ]
      },
      "OperationBase": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "path": {
            "type": "string",
            "nullable": true
          },
          "op": {
            "type": "string",
            "nullable": true
          },
          "from": {
            "type": "string",
            "nullable": true
          }
        }
      }
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "in": "header",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}